Moved all config classes to config package
